Fix dtype for block_usage and block_scales

+++ vllm/worker/hpu_model_runner.py
@@ -1009,7 +1009,7 @@ class HPUModelRunnerBase(ModelRunnerBase[TModelInputForHPU]):
                                      dtype=torch.int,
                                      device=self.device)
         block_usage = torch.tensor(block_usage,
-                                   dtype=torch.bfloat16,
+                                   dtype=self.model_config.dtype,
                                    device=self.device)
 
         slot_mapping = torch.tensor(slot_mapping,
@@ -1018,7 +1018,9 @@ class HPUModelRunnerBase(ModelRunnerBase[TModelInputForHPU]):
 
         block_indices, block_offsets = precompute_indices_and_offsets(
             self.block_size, slot_mapping, False)
-        block_scales = torch.tensor(block_scales, dtype=torch.bfloat16, device=self.device)
+        block_scales = torch.tensor(block_scales,
+                                    dtype=self.model_config.dtype,
+                                    device=self.device)
 
         attn_metadata = self.attn_backend.make_metadata(
             is_prompt=False,
